Split DeploymentTarget.deployments into a module.
I'm about to add another related resource for metadata, so this seemed
like a good idea. There were a number of things at the top of the test
that only pertained to the deployments resource, so that is probably
easier to read now. I almost moved the cert-related items into their own
concern but stopped myself.
